backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "tryadovoi",email = "timaryadovou@yandex.ru"}
]
requires-python = ">=3.11"
dependencies = [
    "pyjwt (>=2.11.0,<3.0.0)",
    "fastapi[standard] (>=0.129.0,<0.130.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)"
]
